fingerprint_study: Add virtual/chromeos-fpmcu-test depend
Add a dependency on chromeos-fpmcu-test, which bring the fputils
library that is optionally used in the fingerprint study.
Although the package is available publicly, the core fputils lib
is only included with the private repo checkout.

Note the package chromeos-fpmcu-test is also used for factory test,
thus the name ends with test.
See b/139063705.

BUG=b:144061947
TEST=# With private repo
     USE=fpstudy ./build_packages --board=hatch
     ./build_image --board=hatch base
     # Check that /opt/fpc/fputils.py in in the image

Change-Id: I5c1795afc6d3ce41dc34a43c59b000c5dde6fe02
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/overlays/chromiumos-overlay/+/2522067
Commit-Queue: Craig Hesling <hesling@chromium.org>
Tested-by: Craig Hesling <hesling@chromium.org>
Reviewed-by: Mike Frysinger <vapier@chromium.org>
Reviewed-by: Tom Hughes <tomhughes@chromium.org>