sys-libs/readline: Moving from portage-stable to chromiumos-overlay.
As part of the changes needed in order to upgrade readline to 8.0 and
mark readline 8.0 as stable, we need to make local changes to the
ebuild.  As a first step, therefore, we need to migrate it from
portage-stable to chromiumos-overlay.

BUG=chromium:1131705
TEST=Successfully 'sudo emerge sys-libs/readline' in chroot.

Change-Id: I69a47d19e6fc05726a165000fab166c5b1bc8ef0
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/overlays/chromiumos-overlay/+/2524768
Auto-Submit: Caroline Tice <cmtice@chromium.org>
Reviewed-by: Manoj Gupta <manojgupta@chromium.org>
Commit-Queue: Caroline Tice <cmtice@chromium.org>
Tested-by: Caroline Tice <cmtice@chromium.org>