media-libs/libcamera: Do not strip IPA binaries
libcamera signs its IPA modules (.so files) after they are built. The
signature is later verified when loading the IPA modules and if they do
not match the IPA is treated as an untrusted module. Portage by default
strips all binaries after the build step and modifies the IPA .so files
in so they fail the signature check.

Portage inject hooks after the post_src_install hook that strips
binaries and creates the package that is installed on target. It is
therefore not possible to generate the IPA module signature for the
stripped modules without also packaging the private key and doing so in
pre_pkg_preinst. Stripping and generating signatures for the IPA .so
files in src_install is not possible as the exact method for stripping
them may differ between the ebuild and the build system hook.

The safest route is to never strip the IPA modules. Instead of
restricting stripping of all libcamera binaries use dostrip to only
disable stripping of the IPA modules. The EAPI needs to be increased to
version 7 to support dostrip.

Change-Id: I8a847b1d1b57f7f75484feac4d8367a82c4b4b71
Signed-off-by: Niklas Söderlund <niklas.soderlund@ragnatech.se>
Signed-off-by: Laurent Pinchart <laurent.pinchart@ideasonboard.com>
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/overlays/chromiumos-overlay/+/2744139
Reviewed-by: Tomasz Figa <tfiga@chromium.org>
Commit-Queue: Tomasz Figa <tfiga@chromium.org>
Tested-by: Tomasz Figa <tfiga@chromium.org>