libva: Add patch for protected content
This adds the upstream patch file from:
https://github.com/intel/libva/pull/457

This contains the protected content interfaces that are needed for
secure playback. We are doing this overall for libva rather than
restricting it to the specific target (tgl) so that the Chrome builders
can have usage of these APIs enabled.

This also adds a dependency on libva-intel-media-driver-pavp, which is
in the partner repository. That ebuild uses patches in a private repo
for building the private PAVP version of iHD. The intended usage is that
when the beanstalk USE flag is set, we will build the right version of
iHD for internal builds, and then for Intel they can build it by
setting the intel_ihd_pavp USE flag.

BUG=b:153111783,b:155507318
TEST=emerge-volteer libva with USE combinations of beanstalk, beanstalk
-internal, beanstalk -internal intel_ihd_pavp, and nothing

Cq-Depend: chrome-internal:3392574,chrome-internal:3401844
Change-Id: Ic2a481165c93a90e16097e980a5479c8cea1076e
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/overlays/chromiumos-overlay/+/2533204
Reviewed-by: Mike Frysinger <vapier@chromium.org>
Reviewed-by: Stéphane Marchesin <marcheu@chromium.org>
Commit-Queue: Jeffrey Kardatzke <jkardatzke@google.com>
Tested-by: Jeffrey Kardatzke <jkardatzke@google.com>