dedede: Add lantis coreboot configuration
Create a new coreboot configuration for the lantis variant
of the waddledee reference board. The new configuration file is
a copy of the reference board, but the name of the baseboard is
replaced by the name of the variant where applicable.

(Auto-Generated by create_coreboot_config.sh version 2.3.0).

BUG=b:171546871
TEST=FW_NAME=lantis emerge-dedede coreboot chromeos-bootimage
Ensure that image-lantis.*.bin are created

Change-Id: If4ecd9e963a2aa6dd9b4fc8bd0138ce7cc55f504
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/overlays/chromiumos-overlay/+/2534531
Reviewed-by: Paul Fagerburg <pfagerburg@chromium.org>
Tested-by: Henry Sun <henrysun@google.com>
Commit-Queue: Paul Fagerburg <pfagerburg@chromium.org>