zork: Add shuboz coreboot configuration
Create a new coreboot configuration for the shuboz variant
of the dalboz reference board. The new configuration file is
a copy of the reference board, but the name of the baseboard is
replaced by the name of the variant where applicable.

(Auto-Generated by create_coreboot_config.sh version 2.3.0).

BUG=b:172021093
TEST=FW_NAME=shuboz emerge-zork coreboot chromeos-bootimage
Ensure that image-shuboz.*.bin are created

Change-Id: Ib0558c7366de0c6c07d722a93fd9d47ed6e6f290
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/overlays/chromiumos-overlay/+/2534530
Reviewed-by: Kane Chen <kane_chen@pegatron.corp-partner.google.com>
Reviewed-by: Kangheui Won <khwon@chromium.org>
Tested-by: Kane Chen <kane_chen@pegatron.corp-partner.google.com>
Commit-Queue: Kangheui Won <khwon@chromium.org>