sys-kernel: Add support for chromeos-5.10
Note: This adds an OWNERS file, as mandated by 'repo upload'.
This OWNERS file is a simple copy from the top level directory.
The submitter has no idea why that is necessary, and deems it
pointless, but it is what it is.

BUG=b:171715474
TEST=repo sync and build chromeos-5.10

Cq-Depend: chrome-internal:3400976
Change-Id: If29b3e6627de6c0a72f5551ddd93ebc4b38a26f5
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/overlays/chromiumos-overlay/+/2536714
Reviewed-by: Mike Frysinger <vapier@chromium.org>
Reviewed-by: Chris McDonald <cjmcdonald@chromium.org>
Tested-by: Guenter Roeck <groeck@chromium.org>
Commit-Queue: Guenter Roeck <groeck@chromium.org>