rmtfs: Fix directory permissions
We want our directory to be created as only accessible by the owner
(0700).

NOTE: we tried to do this in CL:2473159 but it wasn't quite right.
verify_fsg always runs _before_ rmtfs and so the directory was already
created.

We're fixing verify_fsg to also create with the right permissions, but
people still might have old folders.  It's good to add the chmod.

BUG=b:173505923, b:162465221
TEST=w/ all patches, WiFi starts a bit earlier

Change-Id: I6ff14d5d4b40833134c0a7fbc66fdbbbb33e932e
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/overlays/chromiumos-overlay/+/2544914
Reviewed-by: Evan Green <evgreen@chromium.org>
Tested-by: Douglas Anderson <dianders@chromium.org>
Commit-Queue: Douglas Anderson <dianders@chromium.org>