qc-netmgr: Don't start on wifi SKUs
Avoid a painful 60 second delay waiting for the WDS service to come
online, and in fact don't start at all on wifi SKUs.

This heuristic for wifi vs LTE is admittedly brittle. The long term plan
is for this entire daemon to go away. This is supposed to be a short
term workaround until this daemon dies.

BUG=b:173541480
TEST=Manual testing

Change-Id: I6b061d1ec87ff30fff5193ab400426ec17ecdb19
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/overlays/chromiumos-overlay/+/2548203
Tested-by: Evan Green <evgreen@chromium.org>
Reviewed-by: Douglas Anderson <dianders@chromium.org>
Commit-Queue: Evan Green <evgreen@chromium.org>