ml: make ml only depend on libsoda when dlc is enabled.
SODA will be supported only through DLC. Now we want to achieve the
logic that,

1. libsoda is depended on only when both "dlc" and "ondevice_speech"
are both enabled.
2. Even dlc is enabled, if "ondevice_speech" is not enabled, we still
do not support DLC. (This is useful for roll-out and for when some
chromebooks are too low-end to support SODA).

So the BUILD.gn of mlservice can still use "ondevice_speech" to see if
libsoda is enabled.

BUG=b:160550533
TEST=unit test of ml-service

Change-Id: I34f9a1f48678d13f9c527bc6c58331a10d7b7429
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/overlays/chromiumos-overlay/+/2547822
Tested-by: Honglin Yu <honglinyu@chromium.org>
Commit-Queue: Honglin Yu <honglinyu@chromium.org>
Reviewed-by: Andrew Moylan <amoylan@chromium.org>