cros-workon: merge git checkout logic
The git checkout code currently has two loops back to back:
* First it sees if all the checkouts exist from the manifest.
* If they do, it attempts to check them all out.
* If one fails, we note it, and delete the .git/ dir, but we keep
  trying the rest.
* If they all work, great, we're done.
* If any failed, or none existed locally, we try checking them all
  out from the remote git repository.

There's both a bug here and inefficiencies:
* For multiple repos, if all but one exists locally, we don't skip
  trying to use any of those local caches.
* For nested repos, if the parent failed, but a child passed, when
  we fallback to direct git, it fails because the dir isn't empty.

The only way to resolve the latter issue is to keep syncing until
you have an ebuild that is more in sync with the checkouts.

Unify the logic so we only walk the list once.  For each project,
we use the local cache if it exists.  If it doesn't, or it fails,
we fallback to the git eclass for it alone immediately.

BUG=None
TEST=CQ passes

Change-Id: I60412893d95a9d7be2ae56ebb9069c10a0b18cc7
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/overlays/chromiumos-overlay/+/2553245
Tested-by: Mike Frysinger <vapier@chromium.org>
Reviewed-by: Chris McDonald <cjmcdonald@chromium.org>
Commit-Queue: Mike Frysinger <vapier@chromium.org>