llvm: disable assert for DK_Linker diagnostics
CodeGenAction.cpp contained an assert requiring that CurLinkModule is
set to a non-null value for DK_Linker diagnostics. The intent here to
always report the module name when we are reporting such diagnostics.
In practice, CurLinkModule is not always set, so this assert can
trigger. When it does, the compiler crashes and we get neither the
module name nor the diagnostic. This happens even for warnings that
have been suppressed with -Wno-*, so we are failing builds that
otherwise would have succeeded. Long-term, we probably want to ensure
that the module name is always provided, but short term, it seems
strictly better to report the diagnostic without the module name
than to crash the compiler. This change removes the assert and
modifies the code to handle the case where CurLinkModule is not
set.

BUG=chromium:1128437
TEST=llvm-next cq

Change-Id: Ia6731c790654048f1f70c64fea4bb6fc9efb720d
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/overlays/chromiumos-overlay/+/2559199
Reviewed-by: Caroline Tice <cmtice@chromium.org>
Tested-by: Bob Haarman <inglorion@chromium.org>