cros-rust: Improve cros-workon integration and testing.
This change aims to make it possible to reduce the boilerplate of
cros-workon cros-rust ebuilds, while also providing an initial pass at
supporting running unit tests when the target is a different
architecture.

The following ebuilds were converted over to use the new features:
  * chromeos-base/crosh
  * dev-rust/enumn
  * dev-rust/libvda
  * dev-rust/sync
  * dev-rust/sys_util
  * dev-rust/tlsdate_dbus
  * dev-rust/trace-events

chromeos-base/sirenia was used to check for breakage in the cros-workon
cros-rust use case.

dev-rust/bitflags was used to check the cros-rust third party crate case.

dev-rust/crosvm required some changes for the tests to work in the
sandbox.

BUG=chromium:1109258
TEST=FEATURES=test emerge-${BOARD} crosh sirenia tlsdate_dbus
  dev-rust/bitflags crosvm

Change-Id: I1ddc8c1756b8bf12a9faa7b9190aa5a70625059c
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/overlays/chromiumos-overlay/+/2561335
Tested-by: Allen Webb <allenwebb@google.com>
Reviewed-by: Chirantan Ekbote <chirantan@chromium.org>
Auto-Submit: Allen Webb <allenwebb@google.com>