dev-libs/nss: Add patch to prefer writable tokens for trust
Adds a patch to NSS which makes NSS prefer a writable token that already
contains the certificate when looking for a token that should be used to
store trust settings for a certificate.

This avoids the situation where if the token that already has trust
settings is read-only, NSS uses a fall-back logic to just select the
first writable token.

Contains exactly the non-test changes from
https://phabricator.services.mozilla.com/D99840

BUG=chromium:1132030
TEST=modifying trust through chrome://settings/certificates works,
also with built-in certs after CL:2561386 and CL:2567878

Change-Id: I0940cd0de4655020cb9027b33e585022762d423c
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/overlays/chromiumos-overlay/+/2566763
Reviewed-by: Ryan Sleevi <rsleevi@chromium.org>
Reviewed-by: Mike Frysinger <vapier@chromium.org>
Tested-by: Pavol Marko <pmarko@chromium.org>
Commit-Queue: Pavol Marko <pmarko@chromium.org>