dedede: Add sasuke coreboot configuration
Create a new coreboot configuration for the sasuke variant
of the waddledoo reference board. The new configuration file is
a copy of the reference board, but the name of the baseboard is
replaced by the name of the variant where applicable.

(Auto-Generated by create_coreboot_config.sh version 2.3.0).

BUG=b:172104731
TEST=FW_NAME=sasuke emerge-dedede coreboot chromeos-bootimage
Ensure that image-sasuke.*.bin are created

Cq-Depend: chrome-internal:3425366, chromium:2565347, chromium:2567294
Change-Id: I8d0e108e31e4222c56b46b84149b1d90400ed263
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/overlays/chromiumos-overlay/+/2562864
Tested-by: Raymond Chung <raymondchung@ami.corp-partner.google.com>
Reviewed-by: Paul Fagerburg <pfagerburg@chromium.org>
Reviewed-by: Henry Sun <henrysun@google.com>