syzkaller: install syz-db onto the rootfs
Going forward, syzkaller seed corpus will be stored in textual form
along with the tast test. This will allow us to avoid having binary
blobs in the git repository.
The textual seed corpus will need to be packed before use with syzkaller.
Allow syz-db to be built and installed on the rootfs similar to
how syz-manager is, so that the tast wrapper can use it.

BUG=b:173620903
TEST=./build_packages --board=octopus --withtest --cleanbuild
./build_image --board=octopus --noenable_rootfs_verification test

Change-Id: I0d77b8d34eb948617b778ed175b3cefb1e3024de
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/overlays/chromiumos-overlay/+/2568676
Tested-by: Zubin Mithra <zsm@chromium.org>
Reviewed-by: Allen Webb <allenwebb@google.com>
Commit-Queue: Allen Webb <allenwebb@google.com>