dedede: Add storo coreboot configuration
Create a new coreboot configuration for the storo variant
of the waddledee reference board. The new configuration file is
a copy of the reference board, but the name of the baseboard is
replaced by the name of the variant where applicable.

(Auto-Generated by create_coreboot_config.sh version 2.3.0).

BUG=b:174284884
TEST=FW_NAME=storo emerge-dedede coreboot chromeos-bootimage
Ensure that image-storo.*.bin are created

Change-Id: I6503d083aff71e32d85943882fa1a4f5579158ca
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/overlays/chromiumos-overlay/+/2566862
Reviewed-by: Henry Sun <henrysun@google.com>
Reviewed-by: Paul Fagerburg <pfagerburg@chromium.org>
Tested-by: Wentao Qin <qinwentao@huaqin.corp-partner.google.com>
Commit-Queue: Paul Fagerburg <pfagerburg@chromium.org>