volteer: Add drobit coreboot configuration
Create a new coreboot configuration for the drobit variant
of the volteer reference board. The new configuration file is
a copy of the reference board, but the name of the baseboard is
replaced by the name of the variant where applicable.

(Auto-Generated by create_coreboot_config.sh version 2.1.1).

BUG=b:171947885
TEST=FW_NAME=drobit emerge-volteer coreboot chromeos-bootimage
Ensure that image-drobit.*.bin are created

Signed-off-by: FrankChu <frank_chu@pegatron.corp-partner.google.com>
Change-Id: Ief9d8700adff6cd8ae5d8ffce223e18063e05db2
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/overlays/chromiumos-overlay/+/2567297
Reviewed-by: Patrick Georgi <pgeorgi@chromium.org>
Commit-Queue: Zhuohao Lee <zhuohao@chromium.org>