dedede: Add galtic coreboot configuration
Create a new coreboot configuration for the galtic variant
of the dedede reference board. The new configuration file is
a copy of the reference board, but the name of the baseboard is
replaced by the name of the variant where applicable.

(Auto-Generated by create_coreboot_config.sh version 2.1.1).

BUG=b:170913840
TEST=FW_NAME=galtic emerge-dedede coreboot chromeos-bootimage
Ensure that image-galtic.*.bin are created

Signed-off-by: FrankChu <Frank_Chu@pegatron.corp-partner.google.com>
Change-Id: I587a97dec9297e77e43390fb6fee60b4373c0480
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/overlays/chromiumos-overlay/+/2573962
Reviewed-by: Frank Chu <frank_chu@pegatron.corp-partner.google.com>
Reviewed-by: Paul Fagerburg <pfagerburg@chromium.org>
Tested-by: Frank Chu <frank_chu@pegatron.corp-partner.google.com>
Commit-Queue: Paul Fagerburg <pfagerburg@chromium.org>