Revert "vp8 encoder: Workaround to force performance mode  encode"
This reverts commit 7e30d9073fa965d2e4b441e8365817d2ca7ccfb6.

The original commit 7e30d was a driver level hack to trade-off between
the quality and performance to save power.
Now we have legit ways to handle the quality-speed/power tradeoff
from the upper layer using the CL:
https://chromium-review.googlesource.com/c/chromium/src/+/2544368

BUG=b:141147405
TEST=Encode a video sample with appr.tc and ensure there is no power
consumption peek with or with out this patch.

Change-Id: I4eb74435c05db56b91db89fdba89cb1901f0e368
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/overlays/chromiumos-overlay/+/2577806
Reviewed-by: Miguel Casas <mcasas@chromium.org>
Reviewed-by: Hirokazu Honda <hiroh@chromium.org>
Commit-Queue: Sreerenj Balachandran <sreerenj.balachandran@intel.com>
Tested-by: Sreerenj Balachandran <sreerenj.balachandran@intel.com>