volteer: Add copano coreboot configuration
Create a new coreboot configuration for the copano variant
of the volteer reference board. The new configuration file is
a copy of the reference board, but the name of the baseboard is
replaced by the name of the variant where applicable.

(Auto-Generated by create_coreboot_config.sh version 2.1.1).

BUG=b:174413884
TEST=FW_NAME=copano emerge-volteer coreboot chromeos-bootimage
Ensure that image-copano.*.bin are created

Cq-Depend: chrome-internal:3471946
Change-Id: I61743af8d4ae6a7101083b28746eb9412a19e309
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/overlays/chromiumos-overlay/+/2581700
Tested-by: Hao Chou <hao_chou@pegatron.corp-partner.google.com>
Commit-Queue: Zhuohao Lee <zhuohao@chromium.org>
Reviewed-by: Hao Chou <hao_chou@pegatron.corp-partner.google.com>
Reviewed-by: Zhuohao Lee <zhuohao@chromium.org>