llvm: Install llvm-strip as a regular binary
llvm-strip is a symlink to llvm-objcopy and argv[0] is used
to differentiate between the two tools. This does not work
with standalone toolchains since we lose argv[0] and strip
invocations are treated as objcopy. Therefore, install
llvm-strip as a full binary and avoid breaking builds.

The size increase in SDK is around 3.5MB.

BUG=chromium:1151787
TEST=tested with creating and using standalone toolchain locally.

Change-Id: Ic1fc6a1a18aac6594a9bd1a2d908ffeac89dac73
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/overlays/chromiumos-overlay/+/2586212
Tested-by: Manoj Gupta <manojgupta@chromium.org>
Tested-by: Jian Cai <jiancai@google.com>
Reviewed-by: Jian Cai <jiancai@google.com>
Commit-Queue: Jian Cai <jiancai@google.com>