sys-devel/binutils: bump 2.27 -> 2.35.1
The major change is not building from the AOSP binutils git tree
and instead using a tarball with patches in the ebuild.

This opens an opportunity to do many cleanups like dropping old
gold linker patches from AOSP which are not required by CrOS or
dropping the v9999 ebuild and in general having a much clearer
view of what patches are actually required by CrOS.

In addition to forward-porting relevant patches, a new patch was
added for a segfault encountered in DWP which should be sent
upstream.

BUG=chromium:1158321, chromium:1165433, chromium:1163962
TEST=Local builds for eve (x64), kevin (aarch64) and minnie (arm).

Cq-Depend: chromium:2627397, chromium:2635222, chromium:2626659
Change-Id: I2f8d59127685f7c1a58e337eac638e2529802c48
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/overlays/chromiumos-overlay/+/2626657
Tested-by: Manoj Gupta <manojgupta@chromium.org>
Commit-Queue: Manoj Gupta <manojgupta@chromium.org>
Reviewed-by: Manoj Gupta <manojgupta@chromium.org>