stressapptest: Build stressapptest64 for arm64
In this CL, we try to build 2 binaries for arm64.
1. 32-bit stressapptest
2. 64-bit stressapptest64

When stressapptest64 exists on dut and the arch is aarch64, factory
toolkit will use this 64-bit version binary to run test.

However, we still need to preserve the 32-bit version in this case.
Because the lab needs this version. 32-bit minijail can't sandbox a
64-bit process.

BUG=b:175255825
TEST=1. emerge-asurada stressapptest
     2. /build/asurada/usr/bin/stressapptest64 exists
     3. Run stressapptest64 on dut successfully, and it shows 64-bit
     version

Change-Id: I8ff7c4543cb496200496ff5813756961a67846fa
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/overlays/chromiumos-overlay/+/2596618
Reviewed-by: Manoj Gupta <manojgupta@chromium.org>
Reviewed-by: Hung-Te Lin <hungte@chromium.org>
Tested-by: Yilin Yang (kerker) <kerker@chromium.org>
Commit-Queue: Yilin Yang (kerker) <kerker@chromium.org>