sys-boot/coreboot: Split make_coreboot
This split makes make_coreboot easier to digest as it removes the build
target parameter. I also switch to using the $coreboot variable instead
of the $name variable when calling add_fw_blobs. This is a no-op.

BUG=b:112267918
TEST=emerge-grunt coreboot && cbfstool /build/grunt/firmware/aleena/coreboot.rom print -r FW_MAIN_A
Verified additional firmware was added to the image.

Change-Id: I78ac240792d429a52f7d4d9200133fcacc1542d0
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/overlays/chromiumos-overlay/+/2615994
Commit-Queue: Raul E Rangel <rrangel@chromium.org>
Tested-by: Raul E Rangel <rrangel@chromium.org>
Reviewed-by: Simon Glass <sjg@chromium.org>
Reviewed-by: Jack Rosenthal <jrosenth@chromium.org>