sys-boot/coreboot: Don't run while loop in a subshell
In EAPI 6 it was not possible to die in a subshell. This means that the
while was never able to actually die. With EAPI 7 this is no longer a
problem, but lets remove the subshell for the while loop to remove any
surprises.

BUG=b:112267918
TEST=emerge-zork coreboot

Change-Id: Id787c06dfb1808e0f58726f52e3b5246bc205aa3
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/overlays/chromiumos-overlay/+/2615995
Commit-Queue: Raul E Rangel <rrangel@chromium.org>
Tested-by: Raul E Rangel <rrangel@chromium.org>
Reviewed-by: Simon Glass <sjg@chromium.org>
Reviewed-by: Jack Rosenthal <jrosenth@chromium.org>