binutils-2.35.1: add patch to allow partial gold deployment
Chrome needs the DWP tool which is built as part of gold using very
much of the same code as the linker, so we can't outright disable
it but at the same time we don't want to allow the linker to be
deployed because it can be used accidentally by buggy ebuilds.

Thus we add a patch for a flag to only deploy dwp.

BUG=chromium:1163741
TEST=Local builds (x64/arm/aarch64): verified the gold linker
is not packaged anymore, only DWP which is used by Chrome.

Cq-Depend: chromium:2635222, chromium:2626657, chromium:2627397
Change-Id: I0db3173c32f6bb7f15a6076a3cd48f2858320a67
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/overlays/chromiumos-overlay/+/2626659
Tested-by: Manoj Gupta <manojgupta@chromium.org>
Commit-Queue: Manoj Gupta <manojgupta@chromium.org>
Reviewed-by: Manoj Gupta <manojgupta@chromium.org>