dedede: Add sasukette coreboot configuration
Create a new coreboot configuration for the sasukette variant
of the waddledee reference board. The new configuration file is
a copy of the reference board, but the name of the baseboard is
replaced by the name of the variant where applicable.

(Auto-Generated by create_coreboot_config.sh version 2.5.0).

BUG=b:175848514
TEST=FW_NAME=sasukette emerge-dedede coreboot chromeos-bootimage
Ensure that image-sasukette.*.bin are created

Cq-Depend: chromium:2637767
Change-Id: Ia77d63009203fc9545d0e13777ca997b24abc3c3
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/overlays/chromiumos-overlay/+/2628817
Reviewed-by: Paul Fagerburg <pfagerburg@chromium.org>
Reviewed-by: Weimin Wu <wuweimin@huaqin.corp-partner.google.com>
Commit-Queue: Henry Sun <henrysun@google.com>
Tested-by: Henry Sun <henrysun@google.com>