factory_installer: Add a BDEPEND on chromeos-common-script for tests
In <https://crrev.com/c/2622585> the builders are yelling at me in the
test phase.  They say:

 >>> Test phase: chromeos-base/factory_installer-0.0.1-r146
 /build/BRD/usr/sbin/secure-wipe.sh: line 16: /usr/share/misc/chromeos-common.sh: No such file or directory
  * ERROR: chromeos-base/factory_installer-0.0.1-r146::chromiumos failed (test phase):
  *   integration test failed
  *
  * Call stack:
  *               ebuild.sh, line  125:  Called src_test
  *             environment, line 3404:  Called die
  * The specific snippet of code:
  *       tests/secure-wipe.sh || die "integration test failed"

The test phase runs the file:
  src/platform/factory_installer/tests/secure-wipe.sh

Which in turn executes:
  . ${SYSROOT}/usr/sbin/secure-wipe.sh

Which in turn runs:
  . /usr/share/misc/chromeos-common.sh

In other words, we're running a shell script that's intended to run on
the target but it's including a file from the host environment.  This
isn't great, but it shouldn't really hurt since "chromeos-common.sh"
comes from the same source and so should be the same between the host
and the target.  We should be able to fix this with just dependencies.

NOTE: as per tribal knowledge passed down [1], adding this to the
BDEPEND also apparently needs this added to
"virtual/target-chromium-os-sdk".

[1] https://crrev.com/c/2579991/3#message-336f1f87caa0f657f82f9bd97a924e6ba0d4e766

BUG=chromium:1160096
TEST=Builder isn't unhappy anymore

Change-Id: I83c0b2a32a7e337c875d69d836daea422a9fc39f
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/overlays/chromiumos-overlay/+/2628698
Commit-Queue: Douglas Anderson <dianders@chromium.org>
Reviewed-by: Jack Rosenthal <jrosenth@chromium.org>
Reviewed-by: Mike Nichols <mikenichols@chromium.org>
Reviewed-by: Jasmine Chen <lnishan@google.com>
Tested-by: Jack Rosenthal <jrosenth@chromium.org>