cxxbridge-cmd: Create ebuild and add to sdk
Add ebuild for cxxbridge-cmd. This package is a C++ code generator for
integrating `cxx` crates into non-Cargo builds. This is a build time
dependency as it will be used to create a bridge between cxx and Rust.

Add this to the cros sdk so it can be used during the build to generate
bindings between cxx and rust. Also update some other ebuilds so all
dependencies are met and remove unused dependencies (found via
cleanup_crates.py).

BUG=b:176846222
TEST=sudo emerge cxxbridge-cmd and target-os-sdk-post-cross

Change-Id: I1d378b55eceea3e100a7651aedd003d42cb1e6fb
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/overlays/chromiumos-overlay/+/2634240
Tested-by: Abhishek Pandit-Subedi <abhishekpandit@chromium.org>
Reviewed-by: Mike Frysinger <vapier@chromium.org>
Reviewed-by: Chirantan Ekbote <chirantan@chromium.org>
Commit-Queue: Abhishek Pandit-Subedi <abhishekpandit@chromium.org>