avahi: Log address resolution
This information should help to debug the mDNS issue where some Canon
printers initially fail to set up and only succeed after a substantial
delay. We suspect the bug may be caused by avahi failing to resolve a
.local hostname to an ipv4 address, in which case a warning message will
be generated in the system log. If no such warning message is generated,
that would suggest that the mDNS address resolution is working properly
and the bug lies elsewhere.

BUG=b:172217141
TEST=manual

Change-Id: I165d1d7ef7b3828898e0c96d06c22dc238350643
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/overlays/chromiumos-overlay/+/2646932
Tested-by: Pranav Batra <batrapranav@chromium.org>
Reviewed-by: Sean Kau <skau@chromium.org>
Commit-Queue: Sean Kau <skau@chromium.org>