sys-devel/libs: upgrade GCC 4.9 -> 10.2
The biggest change in this upgrade is that the AOSP GCC git tree
is not used anymore and most of the patches present in that tree
have been dropped because they were backports fixed by upstream
or Android-specific fixes.

Because the git tree is not used anymore, the ebuilds use simple
upstream release tarballs with patches applied on top.

BUG=chromium:1167265
TEST=Locally: built for x86_64, arm and aarch64 targets and boot
tested on Eve, Veyron-minnie and Kevin.

Change-Id: I2aede8377aba83c174914e22f0ea0bfca40b56c9
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/overlays/chromiumos-overlay/+/2650153
Reviewed-by: Manoj Gupta <manojgupta@chromium.org>
Tested-by: Manoj Gupta <manojgupta@chromium.org>