chromeos-init: Install lvm configuration file unconditionally
If an lvm-based stateful partition rolls back to a version that
doesn't support it, clobber-state needs to be able to run lvm2
commands to delete and remove the lvm metadata from disk. Install
the lvm conf to allow lvm commands to run in clobber-state.

BUG=b:172344853
TEST=emerge chromeos-init

Disallow-Recycled-Builds: betty-pi-arc-cq
Change-Id: Iabbbfe00ddd4a531c5f304c1e28aabcccdc8620c
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/overlays/chromiumos-overlay/+/2664086
Tested-by: Sarthak Kukreti <sarthakkukreti@chromium.org>
Commit-Queue: Sarthak Kukreti <sarthakkukreti@chromium.org>
Reviewed-by: Gwendal Grignou <gwendal@chromium.org>