arm64, amd64: Mask debug_fission in chromeos-chrome
debug_fission didn't enforce -gsplit-dwarf on arm64 and amd64 with
thinlto until a recent change
https://chromium-review.googlesource.com/c/chromium/src/+/2657848.
Now when debug_fission is in effect on 64-bit targets it causes problems
with debug info.
Technically 64-bit targets don't need debug_fission and it was added
only for performance considerations which may not be an issue with
thinlto.

This is a temporary fix which will be replaced with the fix in ebuild
when uprev is unblocked.

BUG=b:178492639
TEST=emerge-eve chromeos-chrome; emerge-kevin64 chromeos-chrome

Change-Id: I6c2ea57d4ee887ccc4993a1eac14abce36254d59
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/overlays/chromiumos-overlay/+/2665751
Tested-by: Denis Nikitin <denik@chromium.org>
Reviewed-by: Mike Frysinger <vapier@chromium.org>
Reviewed-by: Manoj Gupta <manojgupta@chromium.org>
Commit-Queue: Denis Nikitin <denik@chromium.org>