dedede: Add kracko coreboot configuration
Create a new coreboot configuration for the kracko variant
of the waddledee reference board. The new configuration file is
a copy of the reference board, but the name of the baseboard is
replaced by the name of the variant where applicable.

(Auto-Generated by create_coreboot_config.sh version 2.5.0).

BUG=b:178092096
TEST=FW_NAME=kracko emerge-dedede coreboot chromeos-bootimage
Ensure that image-kracko.*.bin are created

Change-Id: Id259cf718a7022b73fbdc32352decedde209208e
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/overlays/chromiumos-overlay/+/2670839
Reviewed-by: Paul Fagerburg <pfagerburg@chromium.org>
Commit-Queue: Henry Sun <henrysun@google.com>
Tested-by: Henry Sun <henrysun@google.com>