termina_container_tools: Fix lddtree usage
In CL:2640698 we stopped installing some things for borealis, but
inadvertently stopped passing libnss through lddtree. Since lddtree
added the sysroot to the paths, this caused us to start installing
libraries from the chroot instead of the sysroot, which would have the
wrong architecture for ARM devices, and to not install the
dependencies of libnss.

BUG=chromium:1173343
TEST=crostini.Basic on jacuzzi

Disallow-Recycled-Builds: test-failures
Change-Id: I173956306ebe172bf3aa0d2c8346154ade5ab6ec
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/overlays/chromiumos-overlay/+/2672541
Commit-Queue: Fergus Dall <sidereal@google.com>
Tested-by: Fergus Dall <sidereal@google.com>
Reviewed-by: Nicholas Verne <nverne@chromium.org>