libva: Update for libva upstream protected content API
This adds a patch that introduces a transitional state for libva that is
backwards compatible with the downstream protected content API and
forwards compatible with upstream. This will then allow Chrome to switch
to the upstream API and after that is done we can update this patch to
be in line with upstream as well.

BUG=b:174951211
TEST=Builds, Chrome also builds against chroot and protected content
works

Cq-Depend: chrome-internal:3606861,chrome-internal:3606780
Change-Id: I738992882d18b5ae51817a6f9de02243aaff65a1
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/overlays/chromiumos-overlay/+/2687891
Reviewed-by: Andres Calderon Jaramillo <andrescj@chromium.org>
Reviewed-by: Andres Calderon Jaramillo <andrescj@google.com>
Tested-by: Jeffrey Kardatzke <jkardatzke@google.com>
Commit-Queue: Andres Calderon Jaramillo <andrescj@chromium.org>