PS8705: Adjust hash file creation
With the current hash file creation, the hex argument for the version
is interpreted as octal when the first digit is 0, and this creates
issues processing FW versions above 7.  Fix this, and also correct
newins usage to doins.

BUG=b:178678268
TEST=build AP FW and verify hash in CBFS

Change-Id: I1c975d215ffdfae65f2e28f55dd539495cb50c77
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/overlays/chromiumos-overlay/+/2702411
Tested-by: Diana Z <dzigterman@chromium.org>
Auto-Submit: Diana Z <dzigterman@chromium.org>
Reviewed-by: Mike Frysinger <vapier@chromium.org>
Reviewed-by: Aseda Aboagye <aaboagye@chromium.org>
Commit-Queue: Diana Z <dzigterman@chromium.org>