perfetto: create /run/perfetto directory early at startup
By creating this directory early, we prevent the perfetto
SDK from ever falling back on /tmp/ for the producer socket.
While there is not necessarily any security issue with letting
the perfetto SDK look in /tmp, it is not ever used on Chrome OS, so
it is best to avoid the extra edge case.

BUG=b:177664944
TEST=reboot and confirm that /run/perfetto directory exists even if
     traced is not started.

Cq-Depend: chromium:2717195
Change-Id: Ib7247f60ac36f552bdd90f306767366f5933606a
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/overlays/chromiumos-overlay/+/2715731
Reviewed-by: Chinglin Yu <chinglinyu@chromium.org>
Reviewed-by: Mike Frysinger <vapier@chromium.org>
Reviewed-by: Allen Webb <allenwebb@google.com>
Tested-by: John Bates <jbates@chromium.org>
Commit-Queue: John Bates <jbates@chromium.org>