cr50-scripts: cr50-set-board-id: support whitelabel_dev flags
Currently, whitelabel devices have cr50 board ID flags set to 0x3f80,
regardless which phase it is.  This is causing an issue in
factory_install.sh.  Because the factory install shim would treat these
devices as PVT devices, and expect the write protections to be ON in
this case.

By discussion of b/179626571, we decided to have 4 kinds of flags:

  - whitelabel_pvt: 0x3f80
  - whitelabel_dev: 0x3f7f
  - pvt: 0x7f80
  - dev: 0x7f7f

This CL updates cr50-set-board-id.sh to reflect the new settings.

Commands
  - cr50-set-board-id.sh whitelabel_flags
  - cr50-set-board-id.sh whitelabel <board_id>
are now expand to,
  - cr50-set-board-id.sh whitelabel_pvt_flags
  - cr50-set-board-id.sh whitelabel_dev_flags

  - cr50-set-board-id.sh whitelabel_pvt <board_id>
  - cr50-set-board-id.sh whitelabel_dev <board_id>
The old CLI is not supported anymore.

BUG=b:179626571
TEST=manual test

Cq-Depend: chromium:2714662
Change-Id: I1bcb2156d3158a1a9567cd0dda6ce68272dc98eb
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/overlays/chromiumos-overlay/+/2712209
Tested-by: Stimim Chen <stimim@chromium.org>
Reviewed-by: Mary Ruthven <mruthven@chromium.org>
Reviewed-by: Cheng-Han Yang <chenghan@chromium.org>
Commit-Queue: Mary Ruthven <mruthven@chromium.org>