dedede: Add blipper coreboot configuration
Create a new coreboot configuration for the blipper variant
of the waddledee reference board. The new configuration file is
a copy of the reference board, but the name of the baseboard is
replaced by the name of the variant where applicable.

(Auto-Generated by create_coreboot_config.sh version 2.5.0).

BUG=b:179648964
TEST=FW_NAME=blipper emerge-dedede coreboot chromeos-bootimage
Ensure that image-blipper.*.bin are created

Cq-Depend: chromium:2722220
Change-Id: Ic53b8ba71d0a7b329438e168c9ee0e2f2a56e647
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/overlays/chromiumos-overlay/+/2717606
Reviewed-by: Henry Sun <henrysun@google.com>
Reviewed-by: Paul Fagerburg <pfagerburg@chromium.org>
Commit-Queue: Henry Sun <henrysun@google.com>
Tested-by: Henry Sun <henrysun@google.com>