rmtfs: add a stop dependency on shill
rmtfs service is an important service for wifi+modem combo
solution. The wifi firmware is released when the rmtfs stop is
triggered. Currently stop rmtfs is triggered on starting pre-shutdown
and since other dependent service like shill also stops on pre-shutdown.
The firmware is release at the same time when shill stop is triggered,
a race situation is created and some of the shill requests can timeout.
The solution is to add a "stopped shill" dependency on rmtfs.

BUG=b:157657019
TEST=The race situation does not happen any more and kernel warning
message is not seen anymore.

Change-Id: I3b0396e70880b60a055a7994e4bf15ca5870f9fa
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/overlays/chromiumos-overlay/+/2717926
Tested-by: Abhishek Kumar <kuabhs@chromium.org>
Reviewed-by: Brian Norris <briannorris@chromium.org>
Reviewed-by: Douglas Anderson <dianders@chromium.org>
Commit-Queue: Abhishek Kumar <kuabhs@chromium.org>