ti50-sdk: initial import
This imports Ti50's SDK into CrOS, trying as much as possible to keep
with the scripts/setup that Ti50 has been running with so far.

Ti50's SDK is ~2GB, and occupies /opt/ti50-sdk. It contains a mixture of
clang, newlib, and rust. The design of this was discussed here:
https://docs.google.com/document/d/1Fs_oRvanPtEPBia1p88224FVt0Jt1iAr0CDICW_qrY8/edit

It's hoped that, in the future, we'll integrate this all with Chrome OS'
primary C/C++/Rust toolchains. Given the deadlines of various parties,
that doesn't seem feasible to do in our first pass.

BUG=chromium:1179917
TEST=`sudo emerge ti50-sdk`; ti50 builds with a few internal patches.

Change-Id: I8c3e9fa5011b447cc0025b9c0b1ff963bdaad805
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/overlays/chromiumos-overlay/+/2792884
Commit-Queue: George Burgess <gbiv@chromium.org>
Tested-by: George Burgess <gbiv@chromium.org>
Reviewed-by: Mike Frysinger <vapier@chromium.org>