virtual/target-chromium-os: Include chromeos-accelerometer-init only when iioservice is not present
When iioservice is present, it install the conf file trigger udev
(cros-ec-accel.conf and eventually chromeos-ec-driver-init).
chromeos-accelerometer-init is not needed anymore when iioservice is
loaded and will be removed when all chromebook are converted.

BUG=b:169341845
TEST=On volteer with iioservice set, check all needed file are present,
mems_setup for accelerometers start as soon a boot-service is ready.
On snappy with iioservice not set, check all needed file are present,
and mems_setup for accelerometers start before session_manager.

Cq-Depend: chromium:2522505
Change-Id: Ic96d633d50d56f44f0c323e0d499e4d6a2ca531b
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/overlays/chromiumos-overlay/+/2733398
Reviewed-by: Cheng-Hao Yang <chenghaoyang@chromium.org>
Tested-by: Gwendal Grignou <gwendal@chromium.org>
Commit-Queue: Gwendal Grignou <gwendal@chromium.org>