chromeos: do not set v4l2_codec USE flag unconditionally
v4l2_codec is used to compile and use V4L2 video acceleration
on the platform. Not all platforms support this interface, and
those that do are correctly marked via chipset-* or overlay-*
specific USE variables. This CL unsets that variable at the top
level so we don't compile this code and we don't waste time in
Chrome and Chromium trying to interact with this API in platforms
(the majority) that don't support it.

BUG=b:179443596
TEST=CQ Build

Cq-Depend: chromium:2737014
Change-Id: I06e0253c824d9604d60dcacb561fba1055626a4f
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/overlays/chromiumos-overlay/+/2732961
Reviewed-by: Mike Frysinger <vapier@chromium.org>
Reviewed-by: Alex Klein <saklein@chromium.org>
Reviewed-by: Ilja H. Friedel <ihf@chromium.org>
Reviewed-by: Hirokazu Honda <hiroh@chromium.org>
Commit-Queue: Miguel Casas <mcasas@chromium.org>
Tested-by: Miguel Casas <mcasas@chromium.org>