volteer: Add collis coreboot configuration
Create a new coreboot configuration for the collis variant
of the volteer reference board. The new configuration file is
a copy of the reference board, but the name of the baseboard is
replaced by the name of the variant where applicable.

(Auto-Generated by create_coreboot_config.sh version 2.5.0).

BUG=b:182227204
TEST=FW_NAME=collis emerge-volteer coreboot chromeos-bootimage
Ensure that image-collis.*.bin are created

Cq-Depend: chrome-internal:3672904,chrome-internal:3740771
Change-Id: I9534d123c22cfd8524df4860c60da6191c9ed11b
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/overlays/chromiumos-overlay/+/2747080
Reviewed-by: Nick Vaccaro <nvaccaro@google.com>
Reviewed-by: Zhuohao Lee <zhuohao@chromium.org>
Tested-by: Frank Chu <frank_chu@pegatron.corp-partner.google.com>
Commit-Queue: Zhuohao Lee <zhuohao@chromium.org>