libchromeos-use-flags: Remove disable_yuv420_biplanar
This change removes the disable_yuv420_biplanar flag from all libchromeos-use-flags ebuilds. It is one of several CLs to be submitted together for removing the flag.

Chromium/Chrome uses gfx::BufferFormat::YUV_420_BIPLANAR for video pervasively.

In the past, the Tegra-based boards (nyan and derivatives) had problems with this and we added a ChromeOS per-board USE flag disable_yuv420_biplanar to mark this, consuming this in Chrome as a kDisableYuv420Biplanar flag.

All of this is not needed since Nyan/Tegra reached EOL about half way 2020.

BUG=b:182001139
TEST=Build and deploy image to device. Disable rootfs verification.
     On the DUT, add disable_yuv420_biplanar to /etc/ui_use_flags.txt.
     Log in on a test user, open Chrome, and check the Command Line
     flags in chrome://version. disable-yuv420-biplanar is verified
     absent.

Change-Id: Idb6033d935555a71f1886e0b9319c01c593adf39
Cq-Depend: chromium:2765786, chromium:2766167
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/overlays/chromiumos-overlay/+/2766097
Tested-by: Clarissa Garvey <clarissagarvey@chromium.org>
Commit-Queue: Clarissa Garvey <clarissagarvey@chromium.org>
Reviewed-by: Miguel Casas <mcasas@chromium.org>