fwupd: fix handling of pending updates
Do not check for pending updates on devices that we just updated. It is
possible that fwupd would report a pending update on a device that has
already being updated but it is pending activation. This change adds the
a filter to the get-updates operation opposite to the filter in the
preceding update operation.

BUG=b:181048407,b:141546160,b:167681094
TEST=/sbin/initctl emit fwupdtool-update PLUGIN=thunderbolt GUID=e6f2e4ac-f76a-5fc4-b36a-6f9cb548689e

Change-Id: I452ac388311e6eebae268d5dc4a9ceeb5a52540c
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/overlays/chromiumos-overlay/+/2751369
Reviewed-by: Benson Leung <bleung@google.com>
Tested-by: Daniel Campello <campello@chromium.org>
Commit-Queue: Daniel Campello <campello@chromium.org>