media-libs/mesa: add libglvnd ebuild support
This does not yet enable libglvnd by default. That requires:
- USE=libglvnd
- mesa 21.2 (or 9999)

This currently only adds to the 9999 ebuilds, because the current
release versions can not yet support host big GL. To avoid confusion,
this way we get a build error if someone attempts to build with libglvnd
and a non-9999 mesa, but it does mean that when we uprev mesa to 21.2 we
will need to bring over these changes to the release ebuilds.

BUG=b:177358466
TEST=build mesa-9999 with and without libglvnd:
  cros_workon --board=${BOARD} start mesa
  USE=libglvnd emerge-${BOARD} mesa
  emerge-${BOARD} mesa

Change-Id: I030c67b347bd76d12706c16357e5a8f5b41f69ad
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/overlays/chromiumos-overlay/+/2802939
Tested-by: John Bates <jbates@chromium.org>
Reviewed-by: Stéphane Marchesin <marcheu@chromium.org>
Reviewed-by: Matt Turner <msturner@google.com>
Reviewed-by: David Riley <davidriley@chromium.org>
Commit-Queue: John Bates <jbates@chromium.org>